LPS-74373 Move DynamicIncludes to proper package
